Update and rename 2014-2015/2015 February/Cow Hopscotch (Bronze).cpp to 2014-2015/2015 February/bronze/Cow Hopscotch (Bronze).cpp

+++ 2014-2015/2015 February/bronze/Cow Hopscotch (Bronze).cpp
@@ -4,7 +4,7 @@
 #include <map>
 using namespace std;
 #define fast ios_base::sync_with_stdio(false); cin.tie(NULL); cout.tie(NULL);
-typedef int ll;
+typedef long long int ll;
 #define tc ll test;cin >> test;while(test--)
 #define vi vector<ll>
 #define pb push_back
